fix: Amount not update
amount of bank and category not update when delete transaction

+++ frontend/lib/features/home/bloc/transaction/transaction_bloc.dart
@@ -20,11 +20,52 @@ class TransactionBloc extends Bloc<TransactionEvent, TransactionState> {
 
     on<DeleteTransaction>((event, emit) async {
       try {
-        await transactionRepository.deleteTransaction(event.transactionId);
-        add(LoadTransactions());
-        emit(TransactionDeleted());
+        final currentState = state;
+        if (currentState is TransactionsLoaded) {
+          final transaction = currentState.transactions.firstWhere(
+            (t) => t.id == event.transactionId,
+            orElse: () => throw Exception("Transaction not found"),
+          );
+
+          if (transaction == null) {
+            emit(TransactionsLoadError("Transaction not found"));
+            return;
+          }
+
+          final bankId = transaction.bankId;
+          final categoryId = transaction.categoryId;
+          final amount = transaction.amount;
+          final type = transaction.type;
+
+          final bank = await transactionRepository.getBankById(bankId);
+          double newBankAmount = bank.amount;
+
+          if (type == 'income') {
+            newBankAmount -= amount;
+          } else if (type == 'expense') {
+            newBankAmount += amount;
+          }
+
+          await transactionRepository.updateBankAmount(bankId, newBankAmount);
+
+          if (type == 'expense' && categoryId != 0) {
+            final category =
+                await transactionRepository.getCategoryById(categoryId);
+            double newCategoryAmount = category.amount + amount;
+
+            await transactionRepository.updateCategoryAmount(
+                categoryId, newCategoryAmount);
+          }
+
+          await transactionRepository.deleteTransaction(event.transactionId);
+
+          add(LoadTransactions());
+          emit(TransactionDeleted());
+        } else {
+          emit(TransactionsLoadError("Transactions not loaded"));
+        }
       } catch (error) {
-        emit(TransactionsLoading());
+        emit(TransactionsLoadError("Failed to delete transaction: $error"));
       }
     });
   }
